fix bug with py3 where division returns float, not int, un pagination.

+++ lpbm/modules/render.py
@@ -176,9 +176,9 @@ class Render(lpbm.module_loader.Module):
         pwidth = self.modules['config']['paginate.width'] or 5
         pages = int(math.ceil(len(articles) / float(app)))
         def left_stone(page):
-            return min(max(0, page - pwidth / 2), max(0, pages - pwidth))
+            return min(max(0, page - int(pwidth / 2)), max(0, pages - pwidth))
         def right_stone(page):
-            return max(min(pages, page + pwidth / 2), min(pages, pwidth - 1))
+            return max(min(pages, page + int(pwidth / 2)), min(pages, pwidth - 1))
         template = _get_template('articles', 'base.html')
         display_pages = range(1, pages + 1)
         for page in range(pages):
